Setting -> SettingPage, RegisterPage3 created, settingPage connected with TermsInUsePage
