+++ web/src/component/MbtiHistogram/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import ECharts, { EChartsReactProps } from "echarts-for-react";
 import styles from "./index.module.scss";
 import idiotproof from "../../service/idiotproof";
-import Setting from "../Setting";
+import SettingPage from "../SettingPage";
 import { Mbti } from "../../@types/common";
 /*
 [2022.01.12 jongseok lee] 
@@ -10,7 +10,7 @@ MBTI Histogram은 MBTI 변화 추이를 EChart로 그리는 component입니다. 
 만약에 화가님 그림이 더 좋을 경우 컴포넌트 전체 날려버리세요. 구조 파악을 위해 코드는 남겨둡니다.
 */
 export const MBTIHistogram: React.FC = (properties: Properties) => {
-  const id = [`_${idiotproof.trace(Setting)}`, properties.id].join();
+  const id = [`_${idiotproof.trace(MBTIHistogram)}`, properties.id].join();
   const cl = [styles.index, properties.className].join(" ");
 
   // if (!localStorage.getItem("Mbti")) {
