+++ web/src/component/RegisterPage1/index.tsx
@@ -27,10 +27,7 @@ interface Props extends Properties {
 const RegisterPage1 = (properties: Props) => {
   const id = [`_${idiotproof.trace(RegisterPage1)}`, properties.id].join();
   const cl = [styles.index, properties.className].join(" ");
-  let userData: any = localStorage.getItem("userData");
-  if (userData == null || userData.nickname == undefined) {
-    userData = getInitUserData();
-  }
+  const [userData, setUserData] = useState<any>("");
   const [nickname, setNickname] = useState("");
   const [MBTIStates, setMBTIStates] = useState<MBTIStates[]>([
     { MBTI: "E", state: false },
@@ -53,13 +50,18 @@ const RegisterPage1 = (properties: Props) => {
   properties.setNavVisible(false);
 
   useEffect(() => {
-    setNickname(userData.nickname);
-    setMBTIStates(MBTIValueToState(userData.mbtis[0]));
+    console.log("tmp");
+    let tmp: any = localStorage.getItem("userData");
+    if (tmp != null) tmp = JSON.parse(tmp);
+    if (tmp == null || tmp.nickname == undefined) tmp = getInitUserData();
+    setUserData(tmp);
+    setNickname(tmp.nickname);
+    setMBTIStates(MBTIValueToState(tmp.mbtis[0]));
     setBloodTypeStates([
-      { bloodType: "A", state: userData.bloodType == "A" ? true : false },
-      { bloodType: "B", state: userData.bloodType == "B" ? true : false },
-      { bloodType: "AB", state: userData.bloodType == "AB" ? true : false },
-      { bloodType: "O", state: userData.bloodType == "O" ? true : false },
+      { bloodType: "A", state: tmp.bloodType == "A" ? true : false },
+      { bloodType: "B", state: tmp.bloodType == "B" ? true : false },
+      { bloodType: "AB", state: tmp.bloodType == "AB" ? true : false },
+      { bloodType: "O", state: tmp.bloodType == "O" ? true : false },
     ]);
   }, []);
   const handleBackButton = () => {
