+++ web/src/component/TermsInUsePage/index.module.scss
@@ -3,7 +3,6 @@
   padding-left: 25px;
   display: flex;
   flex-direction: column;
-  justify-content: center;
   > .back-button {
     width: 24px;
     height: 24px;
@@ -19,7 +18,7 @@
   }
   > .term-box {
     width: 300px;
-    height: 413px;
+    height: calc(100vh - 300px);
     display: flex;
     flex-direction: column;
     flex: none;
@@ -42,55 +41,5 @@
         color: rgba(4, 2, 2, 0.35);
       }
     }
-  }
-  > .agree_box {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    flex: none;
-    order: 3;
-    margin-top: 17px;
-    > .checkbox_wrapper {
-      display: flex;
-      flex-direction: row;
-      margin-left: 10px;
-      align-items: center;
-      > .check_box {
-        box-sizing: border-box;
-        width: 11px;
-        height: 11px;
-        border: 1px solid #000000;
-      }
-      > .checked {
-        background: #000000;
-      }
-      > .unchecked {
-        background: #ffffff;
-      }
-      > .text {
-        font-family: "Inter";
-        font-style: normal;
-        font-weight: 700;
-        font-size: 12px;
-        line-height: 15px;
-        color: #000000;
-
-        margin-left: 5px;
-      }
-    }
-  }
-  > .next-step-button-on {
-    display: flex;
-    flex: none;
-    order: 4;
-    margin-top: 100px;
-    margin-bottom: 30px;
-  }
-  > .next-step-button-off {
-    display: flex;
-    flex: none;
-    order: 3;
-    margin-top: 100px;
-    margin-bottom: 30px;
   }
 }
