+++ web/src/component/TermsInUsePage/index.tsx
@@ -1,38 +1,18 @@
-import React, { useState } from "react";
 import shepherd from "../../service/shepherd";
 
 import idiotproof from "../../service/idiotproof";
 import styles from "./index.module.scss";
 
 import { terms } from "../../resource/termsInUsePageText";
-
-import MainButton from "../MainButton/MainButton";
 import ArrowLeft from "../../img/Arrow_left.png";
 
-interface Props extends Properties {
-  setNavVisible: Function;
-}
-
-export const TermsInUsePage = (properties: Props) => {
+export const TermsInUsePage = (properties: Properties) => {
   const id = [`_${idiotproof.trace(TermsInUsePage)}`, properties.id].join();
   const cl = [styles.index, properties.className].join(" ");
-  console.log(styles);
-
-  // checkbox state
-  const [isAgreed, setIsAgreed] = useState(false);
-  const checkHandler = () => {
-    setIsAgreed(!isAgreed);
-  };
 
   // navigation
   const handleBackButton = () => {
-    shepherd.whip("test", "RegisterPage2");
-  };
-
-  const gotoNextStep = () => {
-    localStorage.setItem("isOldUser", "true");
-    properties.setNavVisible(true);
-    shepherd.whip("test", "NewsletterPage");
+    shepherd.whip("test", "SettingPage");
   };
 
   return (
@@ -45,24 +25,6 @@ export const TermsInUsePage = (properties: Props) => {
           <div className="terms">{terms}</div>
         </div>
       </div>
-
-      <div className="agree_box">
-        <div className="checkbox_wrapper">
-          <div
-            className={`check_box ${isAgreed ? "checked" : "unchecked"}`}
-            onClick={() => checkHandler()}
-          ></div>
-          <div className="text">
-            본인은 19세 이상이며 이용약관에 동의합니다.
-          </div>
-        </div>
-      </div>
-      <button
-        className={`next-step-button${isAgreed ? "-on" : "-off"}`}
-        onClick={isAgreed ? gotoNextStep : () => {}}
-      >
-        계속
-      </button>
     </div>
   );
 };
